+++ files/usr/lib/safeshield/dns.sh
@@ -1,5 +1,56 @@
 # shellcheck shell=ash
 
+readonly SS_MIN_DNSMASQ_VERSION='2.80'
+ss_dnsmasq_check_error=''
+
+ss_dnsmasq_version() {
+	dnsmasq --version 2>/dev/null | awk '
+		NR == 1 && $1 == "Dnsmasq" && $2 == "version" {
+			version = $3
+			sub(/[^0-9.].*$/, "", version)
+			print version
+			exit
+		}
+	'
+}
+
+ss_require_supported_dnsmasq() {
+	local current_version
+
+	ss_dnsmasq_check_error=''
+	ss_status_set dnsmasq_min_version "${SS_MIN_DNSMASQ_VERSION}"
+
+	if ! check_dnsmasq_binary; then
+		ss_dnsmasq_check_error='dnsmasq_binary_not_found'
+		ss_status_set dnsmasq_version ''
+		ss_status_set health_dnsmasq_binary '0'
+		ss_status_set health_dnsmasq_version '0'
+		log_error 'dnsmasq binary not found'
+		return 1
+	fi
+	ss_status_set health_dnsmasq_binary '1'
+
+	current_version="$(ss_dnsmasq_version)"
+	ss_status_set dnsmasq_version "${current_version}"
+
+	if [ -z "${current_version}" ]; then
+		ss_dnsmasq_check_error='dnsmasq_version_unknown'
+		ss_status_set health_dnsmasq_version '0'
+		log_error 'Unable to determine dnsmasq version'
+		return 1
+	fi
+
+	if ! is_greater_equal "${current_version}" "${SS_MIN_DNSMASQ_VERSION}"; then
+		ss_dnsmasq_check_error='dnsmasq_version_unsupported'
+		ss_status_set health_dnsmasq_version '0'
+		log_error "Unsupported dnsmasq version ${current_version}; SafeShield requires ${SS_MIN_DNSMASQ_VERSION} or later"
+		return 1
+	fi
+
+	ss_status_set health_dnsmasq_version '1'
+	return 0
+}
+
 ss_mkdirs() {
 	mkdir -p "${SS_TMP_DIR}" "${SS_DNSMASQ_DIR}"
 }